+++ otj-metrics-core/src/main/java/com/opentable/metrics/common/CheckController.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.opentable.metrics.ready;
+package com.opentable.metrics.common;
 
 import java.util.Arrays;
 import java.util.Collections;
