+++ otj-metrics-mvc/src/main/java/com/opentable/metrics/mvc/ReadyEndpoint.java
@@ -29,7 +29,7 @@ import com.opentable.metrics.http.CheckState;
 import com.opentable.metrics.ready.ReadyCheck;
 import com.opentable.metrics.ready.ReadyConfiguration;
 import com.opentable.metrics.ready.ReadyController;
-import com.opentable.metrics.ready.SortedEntry;
+import com.opentable.metrics.common.SortedEntry;
 
 @RestController
 @RequestMapping(ReadyConfiguration.READY_CHECK_PATH)
