+++ otj-metrics-reactive/src/main/java/com/opentable/metrics/reactive/ReadyEndpoint.java
@@ -30,7 +30,7 @@ import reactor.core.publisher.Mono;
 import com.opentable.metrics.http.CheckState;
 import com.opentable.metrics.ready.ReadyConfiguration;
 import com.opentable.metrics.ready.ReadyController;
-import com.opentable.metrics.ready.SortedEntry;
+import com.opentable.metrics.common.SortedEntry;
 
 /**
  * Ready endpoint for Reactive HTTP servers.
